Refactor descriptor heap stride and range handling
Clarify parameter names from reservedRange to reservedRangeSize.
Change VKDescriptorRegion stride type from nuint to ulong.
Refactor offset, stride, and size calculations for clarity and alignment.
Update VKDescriptorRegion construction to use ulong stride.
Range property now returns DeviceAddressRangeEXT.
Minor code cleanup and improved variable naming.

+++ sources/Zenith.NET.Vulkan/VKDescriptorRegion.cs
@@ -2,7 +2,7 @@
 
 namespace Zenith.NET.Vulkan;
 
-internal unsafe class VKDescriptorRegion(uint baseIndex, nuint stride)
+internal unsafe class VKDescriptorRegion(uint baseIndex, ulong stride)
 {
     private readonly Lock @lock = new();
     private readonly Stack<uint> recycled = [];
@@ -20,7 +20,7 @@ internal unsafe class VKDescriptorRegion(uint baseIndex, nuint stride)
 
         index += baseIndex;
 
-        target = new((void*)(pointer + (nint)(stride * index)), stride);
+        target = new((void*)(pointer + (nint)(stride * index)), (nuint)stride);
 
         return new(this, index);
     }
